Updated jar and fixed bugs
Updated the JAR to 1.3 (No known bugs at time of release). Fixed a few
bugs with the INIFileWriter: Adding an element to a category would
remove all white space and comments from the file, Adding an element to
a file would delete the rest of the content of the file from every line
below the element.