Fixed README file
The old example for the `INIReader#getCategoryCount()` had an error:
When I was concatenating the "Total number of categories: " and the
total number of categories variable (`totalCategories`), I put the
INIReader variable by mistake. Also, the
`INIFileWriter#addCategory(String)` needed to be updated to the current
version, which puts a new line between the end of a previous category.