Package update, added method, and all errors now print instead of being ignored
Since I no longer own the website http://www.PlatinumCoding.netii.net/,
I changed the package layout to my new website, which is
http://www.PlatinumCoding.netne.net/ . Also, added a new method within
`INIFileWriter`, which is `addComment(String category, Object comment)`,
which adds a comment to the line above the category. If the comment
contains '\n', then the it will result in two lines of comments.